+++ src/bfs/bfs.cpp
@@ -24,7 +24,7 @@ void BFS::init() {
 
 bool BFS::nextComponent() {
     bool found = false;
-    for (uint a = 0; a < n; a++) {
+    for (uint64_t a = 0; a < n; a++) {
         if (color.get(a) == BFS_WHITE) {
             u = a;
             found = true;
@@ -50,7 +50,7 @@ bool BFS::hasGrayNode() {
     }
     return true;
 }
-uint BFS::getGrayNode() {
+uint64_t BFS::getGrayNode() {
     uint64_t r = INVALID;
     try {
         r = isInner.choice();
@@ -61,12 +61,12 @@ uint BFS::getGrayNode() {
             throw NoMoreGrayNodes();
         }
     }
-    return static_cast<uint>(r);
+    return static_cast<uint64_t>(r);
 }
 
 bool BFS::more() { return hasGrayNode(); }
 
-std::pair<uint, uint> BFS::next() {
+std::pair<uint64_t, uint64_t> BFS::next() {
     u = getGrayNode();
     if (color.get(u) == outerGray) {
         uint32_t tmp = innerGray;
@@ -75,8 +75,8 @@ std::pair<uint, uint> BFS::next() {
         std::swap(isInner, isOuter);
         dist++;
     }
-    for (uint k = 0; k < g->deg(u); k++) {
-        uint v = g->head(u, k);
+    for (uint64_t k = 0; k < g->deg(u); k++) {
+        uint64_t v = g->head(u, k);
         preexplore(u, v);
         if (color.get(v) == BFS_WHITE) {
             preprocess(v);
@@ -86,10 +86,10 @@ std::pair<uint, uint> BFS::next() {
     }
     isInner.remove(u);
     color.insert(u, BFS_BLACK);
-    return std::pair<uint, uint>(u, dist);
+    return std::pair<uint64_t, uint64_t>(u, dist);
 }
 
-void BFS::forEach(std::function<void(std::pair<uint, uint>)> f) {
+void BFS::forEach(std::function<void(std::pair<uint64_t, uint64_t>)> f) {
     do {
         while (more()) f(next());
     } while (nextComponent());
@@ -103,7 +103,7 @@ BFS::BFS(Graph const *graph, Consumer pp, BiConsumer pe)
       isOuter(n),
       preprocess(pp),
       preexplore(pe) {
-    for (uint a = 0; a < n; a++) color.insert(a, BFS_WHITE);
+    for (uint64_t a = 0; a < n; a++) color.insert(a, BFS_WHITE);
 }
 
 BFS::BFS(Graph const *graph, CompactArray c, Consumer pp, BiConsumer pe)
@@ -114,7 +114,7 @@ BFS::BFS(Graph const *graph, CompactArray c, Consumer pp, BiConsumer pe)
       isOuter(n),
       preprocess(pp),
       preexplore(pe) {
-    for (uint a = 0; a < n; a++) color.insert(a, BFS_WHITE);
+    for (uint64_t a = 0; a < n; a++) color.insert(a, BFS_WHITE);
 }
 
 }  // namespace Sealib
