+++ src/dfs/dfs.cpp
@@ -6,23 +6,23 @@
 
 namespace Sealib {
 
-void DFS::visit_standard(uint u0, Graph const *g, std::vector<uint> *color,
-                         std::vector<std::pair<uint, uint>> *s,
+void DFS::visit_standard(uint64_t u0, Graph const *g, std::vector<uint64_t> *color,
+                         std::vector<std::pair<uint64_t, uint64_t>> *s,
                          Consumer preprocess, BiConsumer preexplore,
                          BiConsumer postexplore, Consumer postprocess) {
     s->push_back({u0, 0});
     while (!s->empty()) {
-        std::pair<uint, uint> x = s->back();
+        std::pair<uint64_t, uint64_t> x = s->back();
         s->pop_back();
-        uint u = x.first;
-        uint k = x.second;
+        uint64_t u = x.first;
+        uint64_t k = x.second;
         if (color->operator[](u) == DFS_WHITE) {
             preprocess(u);
             color->operator[](u) = DFS_GRAY;
         }
         if (k < g->deg(u)) {
             s->push_back({u, k + 1});
-            uint v = g->head(u, k);
+            uint64_t v = g->head(u, k);
             preexplore(u, v);
             if (color->operator[](v) == DFS_WHITE) {
                 s->push_back({v, 0});
@@ -33,20 +33,20 @@ void DFS::visit_standard(uint u0, Graph const *g, std::vector<uint> *color,
             color->operator[](u) = DFS_BLACK;
             postprocess(u);
             if (u != u0) {
-                uint pu = s->back().first;
+                uint64_t pu = s->back().first;
                 postexplore(pu, u);
             }
         }
     }
 }
 
-void DFS::visit_nloglogn(uint u0, Graph const *g, CompactArray *color,
+void DFS::visit_nloglogn(uint64_t u0, Graph const *g, CompactArray *color,
                          SegmentStack *s,
-                         std::function<void(uint u)> restoration,
+                         std::function<void(uint64_t u)> restoration,
                          Consumer preprocess, BiConsumer preexplore,
                          BiConsumer postexplore, Consumer postprocess) {
-    s->push(std::pair<uint, uint>(u0, 0));
-    std::pair<uint, uint> x;
+    s->push(std::pair<uint64_t, uint64_t>(u0, 0));
+    std::pair<uint64_t, uint64_t> x;
     while (!s->isEmpty()) {
         uint8_t sr = s->pop(&x);
         if (sr == DFS_DO_RESTORE) {
@@ -55,7 +55,7 @@ void DFS::visit_nloglogn(uint u0, Graph const *g, CompactArray *color,
         } else if (sr == DFS_NO_MORE_NODES) {
             return;
         }
-        uint u, k;
+        uint64_t u, k;
         u = x.first;
         k = x.second;
         if (color->get(u) == DFS_WHITE) {
@@ -63,23 +63,23 @@ void DFS::visit_nloglogn(uint u0, Graph const *g, CompactArray *color,
             color->insert(u, DFS_GRAY);
         }
         if (k < g->deg(u)) {
-            s->push(std::pair<uint, uint>(u, k + 1));
-            uint v = g->head(u, k);
+            s->push(std::pair<uint64_t, uint64_t>(u, k + 1));
+            uint64_t v = g->head(u, k);
             preexplore(u, v);
             if (color->get(v) == DFS_WHITE) {
-                s->push(std::pair<uint, uint>(v, 0));
+                s->push(std::pair<uint64_t, uint64_t>(v, 0));
             } else {
                 postexplore(u, v);
             }
         } else {
             if (u != u0) {
-                std::pair<uint, uint> px;
+                std::pair<uint64_t, uint64_t> px;
                 sr = s->pop(&px);
                 if (sr == DFS_DO_RESTORE) {
                     restoration(u0);
                     s->pop(&px);
                 }
-                uint pu = px.first;
+                uint64_t pu = px.first;
                 postexplore(pu, u);
                 s->push(px);
             }
@@ -89,27 +89,27 @@ void DFS::visit_nloglogn(uint u0, Graph const *g, CompactArray *color,
     }
 }
 
-void DFS::restore_full(uint u0, Graph const *g, CompactArray *color,
+void DFS::restore_full(uint64_t u0, Graph const *g, CompactArray *color,
                        BasicSegmentStack *s) {
     s->saveTrailer();
     s->dropAll();
-    for (uint a = 0; a < g->getOrder(); a++) {
+    for (uint64_t a = 0; a < g->getOrder(); a++) {
         if (color->get(a) == DFS_GRAY) {
             color->insert(a, DFS_WHITE);
         }
     }
-    s->push(std::pair<uint, uint>(u0, 0));
-    std::pair<uint, uint> x;
+    s->push(std::pair<uint64_t, uint64_t>(u0, 0));
+    std::pair<uint64_t, uint64_t> x;
     while (!s->isAligned()) {
         s->pop(&x);
-        uint u = x.first, k = x.second;
+        uint64_t u = x.first, k = x.second;
         if (color->get(u) == DFS_WHITE) color->insert(u, DFS_GRAY);
         if (k < g->deg(u)) {
-            s->push(std::pair<uint, uint>(u, k + 1));
+            s->push(std::pair<uint64_t, uint64_t>(u, k + 1));
             if (s->isAligned()) break;
-            uint v = g->head(u, k);
+            uint64_t v = g->head(u, k);
             if (color->get(v) == DFS_WHITE)
-                s->push(std::pair<uint, uint>(v, 0));
+                s->push(std::pair<uint64_t, uint64_t>(v, 0));
         }
     }
 }
@@ -123,29 +123,29 @@ void DFS::restore_full(uint u0, Graph const *g, CompactArray *color,
  *    k: outgoing edge that points to a gray, top-segment node (if c is false, k
  * is an 'enclosing' edge)
  */
-static std::pair<bool, uint> findEdge(const uint u, const uint k,
+static std::pair<bool, uint64_t> findEdge(const uint64_t u, const uint64_t k,
                                       Graph const *g, CompactArray *c,
                                       ExtendedSegmentStack *s) {
-    std::pair<bool, uint> r = std::make_pair(false, static_cast<uint>(-1));
-    for (uint i = k; i < g->deg(u); i++) {
-        uint v = g->head(u, i);
+    std::pair<bool, uint64_t> r = std::make_pair(false, static_cast<uint64_t>(-1));
+    for (uint64_t i = k; i < g->deg(u); i++) {
+        uint64_t v = g->head(u, i);
         if (c->get(v) == DFS_GRAY && s->isInTopSegment(v, true)) {
             r = std::make_pair(true, i);
             break;
         }
     }
     if (!r.first) {
-        std::pair<uint, uint> a;
+        std::pair<uint64_t, uint64_t> a;
         s->getTopTrailer(&a);
         r = std::make_pair(false, a.second - 1);
     }
     return r;
 }
 
-void DFS::restore_top(uint u0, Graph const *g, CompactArray *color,
+void DFS::restore_top(uint64_t u0, Graph const *g, CompactArray *color,
                       ExtendedSegmentStack *s) {
-    std::pair<uint, uint> x;
-    uint u = u0, k = 0;
+    std::pair<uint64_t, uint64_t> x;
+    uint64_t u = u0, k = 0;
     if (s->getRestoreTrailer(&x) == 1) {
         color->insert(u, DFS_WHITE);
     } else {
@@ -154,32 +154,32 @@ void DFS::restore_top(uint u0, Graph const *g, CompactArray *color,
         color->insert(u, DFS_WHITE);
     }
     while (!s->isAligned()) {
-        std::pair<bool, uint> r = findEdge(u, k, g, color, s);
-        uint u1 = u, k1 = r.second;
+        std::pair<bool, uint64_t> r = findEdge(u, k, g, color, s);
+        uint64_t u1 = u, k1 = r.second;
         if (r.first) {
-            s->push(std::pair<uint, uint>(
+            s->push(std::pair<uint64_t, uint64_t>(
                 u1, k1 + 1));  // k+1 to simulate the normal stack behaviour
             u = g->head(u1, k1);
             k = s->getOutgoingEdge(u);
             color->insert(u, DFS_WHITE);
         } else {
-            s->push(std::pair<uint, uint>(u1, k1 + 1));
+            s->push(std::pair<uint64_t, uint64_t>(u1, k1 + 1));
             // restoration loop must end now, the stack is aligned
         }
     }
     s->recolorLow(DFS_GRAY);
 }
 
-void DFS::visit_nplusm(uint u0, UndirectedGraph const *g, CompactArray *color,
+void DFS::visit_nplusm(uint64_t u0, UndirectedGraph const *g, CompactArray *color,
                        Sequence<uint64_t> *back, Consumer preprocess,
                        BiConsumer preexplore, BiConsumer postexplore,
                        Consumer postprocess) {
     color->insert(u0, DFS_GRAY);
     preprocess(u0);
-    uint u = u0, k = 0;
+    uint64_t u = u0, k = 0;
     while (true) {
         if (k < g->deg(u)) {
-            uint v = g->head(u, k);
+            uint64_t v = g->head(u, k);
             preexplore(u, k);
             if (color->get(v) == DFS_WHITE) {
                 preprocess(v);
@@ -196,8 +196,8 @@ void DFS::visit_nplusm(uint u0, UndirectedGraph const *g, CompactArray *color,
             color->insert(u, DFS_BLACK);
             postprocess(u);
             if (u != u0) {
-                uint pk = g->mate(u, static_cast<uint>(back->get(u))),
-                     pu = g->head(u, static_cast<uint>(back->get(u)));
+                uint64_t pk = g->mate(u, static_cast<uint64_t>(back->get(u))),
+                     pu = g->head(u, static_cast<uint64_t>(back->get(u)));
                 postexplore(pu, pk);
                 u = pu;
                 k = pk + 1;
@@ -211,9 +211,9 @@ void DFS::visit_nplusm(uint u0, UndirectedGraph const *g, CompactArray *color,
 void DFS::standardDFS(Graph const *g, Consumer preprocess,
                       BiConsumer preexplore, BiConsumer postexplore,
                       Consumer postprocess) {
-    std::vector<uint> color(g->getOrder());
-    std::vector<std::pair<uint, uint>> s;
-    for (uint u = 0; u < g->getOrder(); u++) {
+    std::vector<uint64_t> color(g->getOrder());
+    std::vector<std::pair<uint64_t, uint64_t>> s;
+    for (uint64_t u = 0; u < g->getOrder(); u++) {
         if (color[u] == DFS_WHITE) {
             visit_standard(u, g, &color, &s, preprocess, preexplore,
                            postexplore, postprocess);
@@ -223,10 +223,10 @@ void DFS::standardDFS(Graph const *g, Consumer preprocess,
 
 void DFS::nBitDFS(Graph const *g, Consumer preprocess, BiConsumer preexplore,
                   BiConsumer postexplore, Consumer postprocess) {
-    uint n = g->getOrder();
+    uint64_t n = g->getOrder();
     double e = 0.2;
     unsigned q = static_cast<unsigned>(ceil(
-        ceil(e / 6 * n) / (8 * sizeof(std::pair<uint, uint>))));  // 2q entries
+        ceil(e / 6 * n) / (8 * sizeof(std::pair<uint64_t, uint64_t>))));  // 2q entries
                                                                   // on S shall
                                                                   // take up at
                                                                   // most (e/3)n
@@ -237,11 +237,11 @@ void DFS::nBitDFS(Graph const *g, Consumer preprocess, BiConsumer preexplore,
     // printf("e=%3.2f, q=%u, n=%u\n", e, q, n);
     BasicSegmentStack s(q);
     CompactArray color(n, 3);
-    for (uint a = 0; a < n; a++) color.insert(a, DFS_WHITE);
-    for (uint a = 0; a < n; a++) {
+    for (uint64_t a = 0; a < n; a++) color.insert(a, DFS_WHITE);
+    for (uint64_t a = 0; a < n; a++) {
         if (color.get(a) == DFS_WHITE)
             visit_nloglogn(a, g, &color, &s,
-                           [&](uint u0) { restore_full(u0, g, &color, &s); },
+                           [&](uint64_t u0) { restore_full(u0, g, &color, &s); },
                            preprocess, preexplore, postexplore, postprocess);
     }
 }
@@ -249,14 +249,14 @@ void DFS::nBitDFS(Graph const *g, Consumer preprocess, BiConsumer preexplore,
 void DFS::nloglognBitDFS(Graph const *g, Consumer preprocess,
                          BiConsumer preexplore, BiConsumer postexplore,
                          Consumer postprocess) {
-    uint n = g->getOrder();
+    uint64_t n = g->getOrder();
     CompactArray color(n, 3);
     ExtendedSegmentStack s(n, g, &color);
-    for (uint a = 0; a < n; a++) color.insert(a, DFS_WHITE);
-    for (uint a = 0; a < n; a++) {
+    for (uint64_t a = 0; a < n; a++) color.insert(a, DFS_WHITE);
+    for (uint64_t a = 0; a < n; a++) {
         if (color.get(a) == DFS_WHITE)
             visit_nloglogn(a, g, &color, &s,
-                           [&](uint u0) { restore_top(u0, g, &color, &s); },
+                           [&](uint64_t u0) { restore_top(u0, g, &color, &s); },
                            preprocess, preexplore, postexplore, postprocess);
     }
 }
@@ -264,26 +264,26 @@ void DFS::nloglognBitDFS(Graph const *g, Consumer preprocess,
 void DFS::nplusmBitDFS(UndirectedGraph const *g, Consumer preprocess,
                        BiConsumer preexplore, BiConsumer postexplore,
                        Consumer postprocess) {
-    uint n = g->getOrder();
+    uint64_t n = g->getOrder();
     CompactArray color(n, 3);
-    for (uint a = 0; a < n; a++) color.insert(a, DFS_WHITE);
+    for (uint64_t a = 0; a < n; a++) color.insert(a, DFS_WHITE);
     std::vector<bool> bits;
-    for (uint u = 0; u < n; u++) {
+    for (uint64_t u = 0; u < n; u++) {
         bits.push_back(1);
-        for (uint k = 0; k < ceil(log2(g->deg(u) + 1)); k++) {
+        for (uint64_t k = 0; k < ceil(log2(g->deg(u) + 1)); k++) {
             bits.push_back(0);
         }
     }
     StaticSpaceStorage back(bits);
-    for (uint a = 0; a < n; a++) {
+    for (uint64_t a = 0; a < n; a++) {
         if (color.get(a) == DFS_WHITE)
             visit_nplusm(a, g, &color, &back, preprocess, preexplore,
                          postexplore, postprocess);
     }
 }
 
-void DFS::runLinearTimeInplaceDFS(uint *graph, Consumer preprocess,
-                                  Consumer postprocess, uint startVertex) {
+void DFS::runLinearTimeInplaceDFS(uint64_t *graph, Consumer preprocess,
+                                  Consumer postprocess, uint64_t startVertex) {
     auto *ilDFSRunner =
         new LinearTimeInplaceDFSRunner(graph, preprocess, postprocess);
     ilDFSRunner->run(startVertex);
